crawl/requirements.txt
KONLP
Mecab
openpyxl
nltk
numpy
pandas
seaborn
tqdm
bs4
requests